+++ CHANGELOG.md
@@ -14,6 +14,7 @@ in case active view was not set
 
 ### Changed
 - Refactored a few classes for easier testability
+- Separated language and number format setting into two settings
 
 ## [0.2.0] - 2018-01-04
 ### Added